move token_authenticator.dart to auth package
